+++ controller/dataplane/owned_finalizer_controller.go
@@ -19,6 +19,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 
 	"github.com/kong/gateway-operator/controller/pkg/log"
+	"github.com/kong/gateway-operator/modules/manager/logging"
 	"github.com/kong/gateway-operator/pkg/consts"
 	k8sutils "github.com/kong/gateway-operator/pkg/utils/kubernetes"
 
@@ -63,8 +64,8 @@ type DeepCopier[T DataPlaneOwnedResource, PT ClientObjectPointer[T]] interface {
 // This is a stop gap solution until we implement proper self-healing for the DataPlane resources, see:
 // https://github.com/Kong/gateway-operator/issues/1028
 type DataPlaneOwnedResourceFinalizerReconciler[T DataPlaneOwnedResource, PT DataPlaneOwnedResourcePointer[T, PT]] struct {
-	Client          client.Client
-	DevelopmentMode bool
+	Client      client.Client
+	LoggingMode logging.Mode
 }
 
 // NewDataPlaneOwnedResourceFinalizerReconciler returns a new DataPlaneOwnedResourceFinalizerReconciler for a type passed
@@ -78,11 +79,11 @@ type DataPlaneOwnedResourceFinalizerReconciler[T DataPlaneOwnedResource, PT Data
 //	NewDataPlaneOwnedResourceFinalizerReconciler[corev1.Service, *corev1.Service](...).
 func NewDataPlaneOwnedResourceFinalizerReconciler[T DataPlaneOwnedResource, PT DataPlaneOwnedResourcePointer[T, PT]](
 	client client.Client,
-	developmentMode bool,
+	loggingMode logging.Mode,
 ) *DataPlaneOwnedResourceFinalizerReconciler[T, PT] {
 	return &DataPlaneOwnedResourceFinalizerReconciler[T, PT]{
-		Client:          client,
-		DevelopmentMode: developmentMode,
+		Client:      client,
+		LoggingMode: loggingMode,
 	}
 }
 
@@ -119,7 +120,7 @@ func (r DataPlaneOwnedResourceFinalizerReconciler[T, PT]) Reconcile(ctx context.
 		return ctrl.Result{}, fmt.Errorf("failed to get %s %s/%s: %w", obj.GetObjectKind().GroupVersionKind().Kind, obj.GetNamespace(), obj.GetName(), err)
 	}
 
-	logger := log.GetLogger(ctx, obj.GetObjectKind().GroupVersionKind().Kind, r.DevelopmentMode)
+	logger := log.GetLogger(ctx, obj.GetObjectKind().GroupVersionKind().Kind, r.LoggingMode)
 
 	// If the object is not being deleted, we don't need to do anything.
 	if obj.GetDeletionTimestamp().IsZero() {