+++ pkg/utils/kubernetes/resources/clusterrole_helpers_test.go
@@ -13,11 +13,11 @@ import (
 
 func TestClusterroleHelpers(t *testing.T) {
 	testCases := []struct {
-		controlplane        string
-		image               string
-		devMode             bool
-		expectedClusterRole func() *rbacv1.ClusterRole
-		expectedError       error
+		controlplane              string
+		image                     string
+		validateControlPlaneImage bool
+		expectedClusterRole       func() *rbacv1.ClusterRole
+		expectedError             error
 	}{
 		{
 			controlplane: "test_3.1.2",
@@ -29,9 +29,9 @@ func TestClusterroleHelpers(t *testing.T) {
 			},
 		},
 		{
-			controlplane: "test_3.1_dev",
-			image:        "kong/kubernetes-ingress-controller:3.1",
-			devMode:      true,
+			controlplane:              "test_3.1_dev",
+			image:                     "kong/kubernetes-ingress-controller:3.1",
+			validateControlPlaneImage: false,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
 				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_4("test_3.1_dev")
 				k8sresources.LabelObjectAsControlPlaneManaged(cr)
@@ -39,14 +39,15 @@ func TestClusterroleHelpers(t *testing.T) {
 			},
 		},
 		{
-			controlplane:  "test_3.0",
-			image:         "kong/kubernetes-ingress-controller:3.0.0",
-			expectedError: k8sresources.ErrControlPlaneVersionNotSupported,
+			controlplane:              "test_3.0",
+			image:                     "kong/kubernetes-ingress-controller:3.0.0",
+			validateControlPlaneImage: true,
+			expectedError:             k8sresources.ErrControlPlaneVersionNotSupported,
 		},
 		{
-			controlplane: "test_3.0_dev",
-			image:        "kong/kubernetes-ingress-controller:3.0.0",
-			devMode:      true,
+			controlplane:              "test_3.0_dev",
+			image:                     "kong/kubernetes-ingress-controller:3.0.0",
+			validateControlPlaneImage: false,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
 				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_4("test_3.0_dev")
 				k8sresources.LabelObjectAsControlPlaneManaged(cr)
@@ -54,14 +55,15 @@ func TestClusterroleHelpers(t *testing.T) {
 			},
 		},
 		{
-			controlplane:  "test_unsupported",
-			image:         "kong/kubernetes-ingress-controller:1.0",
-			expectedError: k8sresources.ErrControlPlaneVersionNotSupported,
+			controlplane:              "test_unsupported",
+			image:                     "kong/kubernetes-ingress-controller:1.0",
+			validateControlPlaneImage: true,
+			expectedError:             k8sresources.ErrControlPlaneVersionNotSupported,
 		},
 		{
-			controlplane: "test_unsupported_dev",
-			image:        "kong/kubernetes-ingress-controller:1.0",
-			devMode:      true,
+			controlplane:              "test_unsupported_dev",
+			image:                     "kong/kubernetes-ingress-controller:1.0",
+			validateControlPlaneImage: false,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
 				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_4("test_unsupported_dev")
 				k8sresources.LabelObjectAsControlPlaneManaged(cr)
@@ -69,14 +71,15 @@ func TestClusterroleHelpers(t *testing.T) {
 			},
 		},
 		{
-			controlplane:  "test_invalid_tag",
-			image:         "test/development:main",
-			expectedError: kgoerrors.ErrInvalidSemverVersion,
+			controlplane:              "test_invalid_tag",
+			image:                     "test/development:main",
+			validateControlPlaneImage: true,
+			expectedError:             kgoerrors.ErrInvalidSemverVersion,
 		},
 		{
-			controlplane: "test_invalid_tag_dev",
-			image:        "test/development:main",
-			devMode:      true,
+			controlplane:              "test_invalid_tag_dev",
+			image:                     "test/development:main",
+			validateControlPlaneImage: false,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
 				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_4("test_invalid_tag_dev")
 				k8sresources.LabelObjectAsControlPlaneManaged(cr)
@@ -84,9 +87,9 @@ func TestClusterroleHelpers(t *testing.T) {
 			},
 		},
 		{
-			controlplane: "cp-3-2-0",
-			image:        "kong/kubernetes-ingress-controller:3.2.0",
-			devMode:      false,
+			controlplane:              "cp-3-2-0",
+			image:                     "kong/kubernetes-ingress-controller:3.2.0",
+			validateControlPlaneImage: true,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
 				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_2_lt3_3("cp-3-2-0")
 				k8sresources.LabelObjectAsControlPlaneManaged(cr)
@@ -94,9 +97,9 @@ func TestClusterroleHelpers(t *testing.T) {
 			},
 		},
 		{
-			controlplane: "cp-3-3-0",
-			image:        "kong/kubernetes-ingress-controller:3.3.0",
-			devMode:      false,
+			controlplane:              "cp-3-3-0",
+			image:                     "kong/kubernetes-ingress-controller:3.3.0",
+			validateControlPlaneImage: true,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
 				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_3_lt3_4("cp-3-3-0")
 				k8sresources.LabelObjectAsControlPlaneManaged(cr)
@@ -104,9 +107,9 @@ func TestClusterroleHelpers(t *testing.T) {
 			},
 		},
 		{
-			controlplane: "cp-3-4-1",
-			image:        "kong/kubernetes-ingress-controller:3.4.1",
-			devMode:      false,
+			controlplane:              "cp-3-4-1",
+			image:                     "kong/kubernetes-ingress-controller:3.4.1",
+			validateControlPlaneImage: true,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
 				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_4("cp-3-4-1")
 				k8sresources.LabelObjectAsControlPlaneManaged(cr)
@@ -117,7 +120,7 @@ func TestClusterroleHelpers(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.controlplane, func(t *testing.T) {
-			clusterRole, err := k8sresources.GenerateNewClusterRoleForControlPlane(tc.controlplane, tc.image, tc.devMode)
+			clusterRole, err := k8sresources.GenerateNewClusterRoleForControlPlane(tc.controlplane, tc.image, tc.validateControlPlaneImage)
 			if tc.expectedError != nil {
 				require.Error(t, err)
 				require.ErrorIs(t, err, tc.expectedError)