+++ test/envtest/kongpluginbinding_managed_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/kong/gateway-operator/controller/konnect"
 	sdkmocks "github.com/kong/gateway-operator/controller/konnect/ops/sdk/mocks"
 	"github.com/kong/gateway-operator/internal/utils/index"
+	"github.com/kong/gateway-operator/modules/manager/logging"
 	"github.com/kong/gateway-operator/modules/manager/scheme"
 	"github.com/kong/gateway-operator/pkg/consts"
 	"github.com/kong/gateway-operator/test/helpers/deploy"
@@ -59,8 +60,8 @@ func TestKongPluginBindingManaged(t *testing.T) {
 	sdk := factory.SDK
 
 	reconcilers := []Reconciler{
-		konnect.NewKongPluginReconciler(false, mgr.GetClient()),
-		konnect.NewKonnectEntityReconciler(factory, false, mgr.GetClient(),
+		konnect.NewKongPluginReconciler(logging.DevelopmentMode, mgr.GetClient()),
+		konnect.NewKonnectEntityReconciler(factory, logging.DevelopmentMode, mgr.GetClient(),
 			konnect.WithKonnectEntitySyncPeriod[configurationv1alpha1.KongPluginBinding](konnectInfiniteSyncTime),
 		),
 	}